+++ Lib/fontbakery/profiles/opentype.py
@@ -78,6 +78,7 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/layout_valid_feature_tags',
     'com.google.fonts/check/layout_valid_script_tags',
     'com.google.fonts/check/layout_valid_language_tags',
+    'com.adobe.fonts/check/varfont/valid_axis_nameid',
 ]
 
 profile.auto_register(globals())